feat:提交

+++ lib_common/src/main/java/com/yang/lib_common/util/AppUtils.kt
@@ -2,12 +2,10 @@
 
 package com.yang.lib_common.util
 
-import android.app.WallpaperManager
-import android.content.ComponentName
+import android.Manifest
+import android.content.ContentValues
 import android.content.Context
-import android.content.Intent
 import android.graphics.Bitmap
-import android.graphics.BitmapFactory
 import android.net.Uri
 import android.os.Build
 import android.os.Environment
@@ -18,7 +16,7 @@ import android.view.View
 import android.view.inputmethod.InputMethodManager
 import androidx.core.app.ActivityOptionsCompat
 import androidx.core.content.FileProvider
-import com.blankj.utilcode.util.RomUtils
+import com.blankj.utilcode.util.*
 import com.google.gson.Gson
 import com.jakewharton.rxbinding4.view.clicks
 import com.tencent.mmkv.MMKV
@@ -27,9 +25,7 @@ import com.yang.lib_common.constant.AppConstant
 import com.yang.lib_common.constant.AppConstant.Constant.CLICK_TIME
 import com.yang.lib_common.room.entity.UserInfoData
 import io.reactivex.rxjava3.core.Observable
-import java.io.File
-import java.io.FileInputStream
-import java.io.FileOutputStream
+import java.io.*
 import java.text.DecimalFormat
 import java.text.SimpleDateFormat
 import java.util.*
@@ -336,22 +332,22 @@ fun formatSize(size: Long): String {
 /**
  * @return 删除文件夹
  */
-fun deleteDirectory(file: File) {
+fun deleteDirectory(file: File,context: Context) {
     try {
         if (file.isDirectory) {
             file.listFiles()?.let {
                 if (it.isNotEmpty()) {
                     for (mFile in it) {
                         if (mFile.isDirectory) {
-                            deleteDirectory(file)
+                            deleteDirectory(file,context)
                         } else {
-                            deleteFile(mFile)
+                            toDeleteFile(mFile,context)
                         }
                     }
                 }
             }
         } else {
-            deleteFile(file)
+            toDeleteFile(file,context)
         }
     }catch (e:Exception){
         e.printStackTrace()
@@ -362,9 +358,15 @@ fun deleteDirectory(file: File) {
 /**
  * @return 删除文件
  */
-fun deleteFile(file: File) {
-    if (file.exists()) {
-        file.delete()
+fun toDeleteFile(file: File, context: Context) {
+    val uri = MediaStore.Images.Media.EXTERNAL_CONTENT_URI
+    val contentResolver = context.contentResolver
+    val url = MediaStore.Images.Media.DATA + "=?"
+    val delete = contentResolver.delete(uri, url, arrayOf(file.absolutePath))
+    if (delete == 0){
+        if (file.exists()) {
+            file.delete()
+        }
     }
 }
 
@@ -460,5 +462,84 @@ fun getUriWithPath(context: Context, filePath: String): Uri {
         Uri.fromFile(File(filePath))
     }
 }
+
+
+fun save2Album(
+    src: Bitmap,
+    dirName: String?,
+    format: Bitmap.CompressFormat,
+    quality: Int,
+    recycle: Boolean,
+    mFileName:String = ""
+): File? {
+    val safeDirName =
+        if (TextUtils.isEmpty(dirName)) Utils.getApp().packageName else dirName!!
+    val suffix = if (Bitmap.CompressFormat.JPEG == format) "JPG" else format.name
+    var fileName = System.currentTimeMillis().toString() + "_" + quality + "." + suffix
+    if (mFileName.isNotEmpty()){
+        fileName = mFileName
+    }
+    return if (Build.VERSION.SDK_INT < Build.VERSION_CODES.Q) {
+        if (!PermissionUtils.isGranted(Manifest.permission.WRITE_EXTERNAL_STORAGE)) {
+            Log.e("ImageUtils", "save to album need storage permission")
+            return null
+        }
+        val picDir =
+            Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_DCIM)
+        val destFile = File(picDir, "$safeDirName/$fileName")
+        if (destFile.exists()){
+            return destFile
+        }
+        if (!ImageUtils.save(src, destFile, format, quality, recycle)) {
+            return null
+        }
+        FileUtils.notifySystemToScan(destFile)
+        destFile
+    } else {
+
+        val file = File("/storage/emulated/0/${Environment.DIRECTORY_DCIM}/$safeDirName/$fileName")
+        if (file.exists()){
+            return file
+        }
+
+        val contentValues = ContentValues()
+        contentValues.put(MediaStore.Images.Media.DISPLAY_NAME, fileName)
+        contentValues.put(MediaStore.Images.Media.MIME_TYPE, "image/*")
+        val contentUri: Uri = if (Environment.getExternalStorageState() == Environment.MEDIA_MOUNTED) {
+            MediaStore.Images.Media.EXTERNAL_CONTENT_URI
+        } else {
+            MediaStore.Images.Media.INTERNAL_CONTENT_URI
+        }
+        contentValues.put(
+            MediaStore.Images.Media.RELATIVE_PATH,
+            Environment.DIRECTORY_DCIM + "/" + safeDirName
+        )
+        contentValues.put(MediaStore.MediaColumns.IS_PENDING, 1)
+        val uri = Utils.getApp().contentResolver.insert(contentUri, contentValues)
+            ?: return null
+        var os: OutputStream? = null
+        try {
+            os = Utils.getApp().contentResolver.openOutputStream(uri)
+            src.compress(format, quality, os)
+            os?.flush()
+            contentValues.clear()
+            contentValues.put(MediaStore.MediaColumns.IS_PENDING, 0)
+            Utils.getApp().contentResolver.update(uri, contentValues, null, null)
+            UriUtils.uri2File(uri)
+        } catch (e: java.lang.Exception) {
+            Utils.getApp().contentResolver.delete(uri, null, null)
+            e.printStackTrace()
+            null
+        } finally {
+            try {
+                os?.close()
+            } catch (e: IOException) {
+                e.printStackTrace()
+            }
+        }
+    }
+}
+
+
